requirements.txt
requests>=2.20.0
azure-functions
